+++ rbx_binary/src/serializer/state.rs
@@ -1,9 +1,8 @@
 use std::{
     borrow::{Borrow, Cow},
-    collections::{BTreeMap, BTreeSet, HashMap, HashSet, VecDeque},
+    collections::{BTreeMap, BTreeSet, HashMap, HashSet},
     convert::TryInto,
     io::Write,
-    u32,
 };
 
 use rbx_dom_weak::{
@@ -248,19 +247,52 @@ impl<'dom, 'db, W: Write> SerializerState<'dom, 'db, W> {
     /// serialization with this serializer.
     #[profiling::function]
     pub fn add_instances(&mut self, referents: &[Ref]) -> Result<(), InnerError> {
-        let mut to_visit = VecDeque::new();
-        to_visit.extend(referents);
+        // Populate relevant_instances with a depth-first post-order traversal over the
+        // tree(s). This is important to ensure that the order of the PRNT chunk (later
+        // written by SerializerState::serialize_parents) is correct.
 
-        while let Some(referent) = to_visit.pop_front() {
+        // The implementation here slightly deviates from Roblox. Roblox writes the PRNT
+        // in depth-first post-order, numbers referents in depth-first pre-order, and
+        // generates type infos in lexical order by class name. See
+        // https://github.com/rojo-rbx/rbx-dom/pull/411#issuecomment-2103713517
+
+        // Since it seems only the PRNT chunk has important semantics related to its
+        // ordering, we do one tree traversal in this function, thereby numbering
+        // referents, generating type infos, and writing the PRNT chunk all in depth-first
+        // post-order.
+        let mut to_visit = Vec::new();
+        let mut last_visited_child = None;
+
+        to_visit.extend(referents.iter().rev());
+
+        while let Some(referent) = to_visit.last() {
             let instance = self
                 .dom
-                .get_by_ref(referent)
-                .ok_or(InnerError::InvalidInstanceId { referent })?;
+                .get_by_ref(*referent)
+                .ok_or(InnerError::InvalidInstanceId {
+                    referent: *referent,
+                })?;
 
-            self.relevant_instances.push(referent);
-            self.collect_type_info(instance)?;
+            to_visit.extend(instance.children().iter().rev());
 
-            to_visit.extend(instance.children());
+            while let Some(referent) = to_visit.last() {
+                let instance =
+                    self.dom
+                        .get_by_ref(*referent)
+                        .ok_or(InnerError::InvalidInstanceId {
+                            referent: *referent,
+                        })?;
+
+                if !instance.children().is_empty()
+                    && instance.children().last() != last_visited_child.as_ref()
+                {
+                    break;
+                }
+
+                self.relevant_instances.push(*referent);
+                self.collect_type_info(instance)?;
+                last_visited_child = to_visit.pop();
+            }
         }
 
         // Sort shared_strings by their hash, to ensure they are deterministically added
@@ -398,9 +430,8 @@ impl<'dom, 'db, W: Write> SerializerState<'dom, 'db, W> {
                 let default_value = type_info
                     .class_descriptor
                     .and_then(|class| {
-                        class
-                            .default_properties
-                            .get(&canonical_name)
+                        database
+                            .find_default_property(class, &canonical_name)
                             .map(Cow::Borrowed)
                     })
                     .or_else(|| Self::fallback_default_value(serialized_ty).map(Cow::Owned))
@@ -800,7 +831,7 @@ impl<'dom, 'db, W: Write> SerializerState<'dom, 'db, W> {
                                 chunk.write_le_u16(value.weight.as_u16())?;
                                 chunk.write_u8(value.style.as_u8())?;
                                 chunk.write_string(
-                                    &value.cached_face_id.clone().unwrap_or_default(),
+                                    value.cached_face_id.as_deref().unwrap_or_default(),
                                 )?;
                             } else {
                                 return type_mismatch(i, &rbx_value, "Font");