+++ rbx_reflection/src/database.rs
@@ -39,6 +39,46 @@ impl<'a> ReflectionDatabase<'a> {
             enums: HashMap::new(),
         }
     }
+
+    /// Returns a list of superclasses for the provided ClassDescriptor. This
+    /// list will start with the provided class and end with `Instance`.
+    pub fn superclasses(
+        &'a self,
+        descriptor: &'a ClassDescriptor<'a>,
+    ) -> Option<Vec<&ClassDescriptor>> {
+        // As of the time of writing (14 March 2024), the class with the most
+        // superclasses has 6 of them.
+        let mut list = Vec::with_capacity(6);
+        let mut current_class = Some(descriptor);
+
+        while let Some(class) = current_class {
+            list.push(class);
+            current_class = class.superclass.as_ref().and_then(|s| self.classes.get(s));
+        }
+
+        Some(list)
+    }
+
+    /// Finds the default value of a property given its name and a class that
+    /// contains or inherits the property. Returns `Some(&Variant)` if a default
+    /// value exists, None otherwise.
+    pub fn find_default_property(
+        &'a self,
+        mut class: &'a ClassDescriptor<'a>,
+        property_name: &str,
+    ) -> Option<&'a Variant> {
+        loop {
+            match class.default_properties.get(property_name) {
+                None => {
+                    class = self
+                        .classes
+                        .get(class.superclass.as_ref()?)
+                        .expect("superclass that is Some should exist in reflection database")
+                }
+                default_value => return default_value,
+            }
+        }
+    }
 }
 
 /// Describes a class of Instance, its properties, and its relation to other