+++ rbx_reflector/src/cli/defaults_place.rs
@@ -2,7 +2,7 @@ use std::{
     fmt::{self, Write},
     fs,
     path::PathBuf,
-    process::Command,
+    process::{Command, Stdio},
     sync::mpsc,
     time::Duration,
 };
@@ -25,7 +25,7 @@ static PLUGIN_SOURCE: &str = include_str!("../../plugin.lua");
 #[derive(Debug, Parser)]
 pub struct DefaultsPlaceSubcommand {
     /// The path of an API dump that came from the dump command.
-    #[clap(long = "api_dump")]
+    #[clap(long)]
     pub api_dump: PathBuf,
     /// Where to output the place. The extension must be .rbxlx
     pub output: PathBuf,
@@ -63,6 +63,8 @@ fn save_place_in_studio(path: &PathBuf) -> anyhow::Result<StudioInfo> {
     log::info!("Starting Roblox Studio...");
 
     let mut studio_process = Command::new(studio_install.application_path())
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
         .arg(path)
         .spawn()?;
 
@@ -85,11 +87,29 @@ fn save_place_in_studio(path: &PathBuf) -> anyhow::Result<StudioInfo> {
         }
     }
 
-    #[cfg(not(target_os = "windows"))]
+    #[cfg(target_os = "macos")]
+    {
+        let process_id = studio_process.id();
+        let script = format!(
+            r#"
+tell application "System Events"
+    set frontmost of the first process whose unix id is {process_id} to true
+    keystroke "s" using command down
+end tell
+"#
+        );
+
+        Command::new("osascript")
+            .args(["-e", script.as_str()])
+            .output()?;
+    }
+
+    #[cfg(not(any(target_os = "windows", target_os = "macos")))]
     println!("Please save the opened place in Roblox Studio (ctrl+s).");
 
     loop {
-        if rx.recv()??.kind.is_create() {
+        let event = rx.recv()??;
+        if event.kind.is_create() || event.kind.is_modify() {
             break;
         }
     }
@@ -127,7 +147,8 @@ fn generate_place_with_all_classes(path: &PathBuf, dump: &Dump) -> anyhow::Resul
             | "Bone"
             | "BaseWrap"
             | "WrapLayer"
-            | "WrapTarget" => continue,
+            | "WrapTarget"
+            | "WrapDeformer" => continue,
 
             "StarterPlayer" => {
                 instance.add_child(Instance::new("StarterPlayerScripts"));
@@ -148,6 +169,7 @@ fn generate_place_with_all_classes(path: &PathBuf, dump: &Dump) -> anyhow::Resul
                 instance.add_child(Instance::new("BaseWrap"));
                 instance.add_child(Instance::new("WrapLayer"));
                 instance.add_child(Instance::new("WrapTarget"));
+                instance.add_child(Instance::new("WrapDeformer"));
             }
 
             _ => {}