+++ rbx_xml/src/tests/models.rs
@@ -66,4 +66,5 @@ model_tests! {
     folder_with_cframe_attributes,
     folder_with_font_attribute,
     number_values_with_security_capabilities,
+    lighting_with_int32_attribute,
 }